Refactor into src screens and utils
